feat: add capability to cli for git comment publishing (#64)
Adds the capability to make a comment to a PR. This was previously being done as part of a CI script, but can now be run as a cli command. We don't want to spam PRs so are making sure that there is only one dextrose comment per PR at any time.